Simplification of the project creation process.
Each time a package is included, the filters need to reset (if any were
set).
Packages currently do not follow any filters set. Does this even matter?
If it does, maybe add some extra parameter to the inclusion functions to
define the systems they are added on. Otherwise, opt for a simpler system
of just creating a whole project with a single table with all packages
needed, extra projects, etc.